requirements.txt
beautifulsoup4
Flask
Flask-WTF
numpy
pandas
werkzeug==0.16.0
urllib3== 1.21.1
flask_bootstrap
requests==2.24.0
mysql-connector-python
idna==2.5
urllib3==1.21.1
requests==2.24.0